+++ packages/elements/src/MatchersCache.ts
@@ -31,7 +31,7 @@ export class MatchersCache extends CacheManager<
   }): string {
     const configHash = `${config.legacyTemplates}|${config.includePaths}|${config.ignorePaths}|${
       config.cache
-    }|${config.multiMatch}|${config.elementDescriptorsPriority}|${config.rootPath}|${config.flagAsExternal.inNodeModules}|${config.flagAsExternal.unresolvableAlias}|${
+    }|${config.descriptorsMultiMatch}|${config.descriptorsPriority}|${config.rootPath}|${config.flagAsExternal.inNodeModules}|${config.flagAsExternal.unresolvableAlias}|${
       config.flagAsExternal.outsideRootPath
     }|${config.flagAsExternal.customSourcePatterns.join(",")}`;